code cleanups
