+++ sample/src/main/java/com/smileidentity/sample/compose/RootScreen.kt
@@ -1,18 +1,8 @@
 package com.smileidentity.sample.compose
 
 import android.content.Context
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.ClickableText
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
+import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -21,17 +11,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalUriHandler
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.chuckerteam.chucker.api.ChuckerInterceptor
@@ -39,8 +19,6 @@ import com.google.mlkit.vision.barcode.common.Barcode
 import com.google.mlkit.vision.codescanner.GmsBarcodeScannerOptions
 import com.google.mlkit.vision.codescanner.GmsBarcodeScanning
 import com.smileidentity.SmileID
-import com.smileidentity.compose.components.BottomPinnedColumn
-import com.smileidentity.compose.components.annotatedStringResource
 import com.smileidentity.sample.BuildConfig
 import com.smileidentity.sample.R
 import com.smileidentity.sample.SmileIDApplication
@@ -58,14 +36,7 @@ import com.smileidentity.viewmodel.viewModelFactory
  * For instructions on initializing the SDK, please refer to [SmileIDApplication].
  */
 @Composable
-fun RootScreen() {
-    SmileIDTheme {
-        Content()
-    }
-}
-
-@Composable
-private fun Content(
+fun RootScreen(
     modifier: Modifier = Modifier,
     viewModel: RootViewModel = viewModel(
         factory = viewModelFactory { RootViewModel() },
@@ -86,7 +57,6 @@ private fun Content(
             .enableAutoZoom()
             .build()
     }
-    val uriHandler = LocalUriHandler.current
     val scanner = remember(context) { GmsBarcodeScanning.getClient(context, options) }
     val client = remember {
         SmileID.getOkHttpClientBuilder()
@@ -94,103 +64,22 @@ private fun Content(
             .build()
     }
 
-    Scaffold(
-        modifier = modifier
-            .fillMaxSize(),
-    ) { paddingValues ->
-        BottomPinnedColumn(
-            modifier = Modifier.padding(paddingValues),
-            scrollableContent = {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_smile_logo),
-                    contentDescription = "",
-                )
-                Text(
-                    text = stringResource(id = R.string.root_welcome),
-                    modifier = Modifier.padding(horizontal = 16.dp),
-                    style = MaterialTheme.typography.bodyLarge.copy(
-                        fontSize = 32.sp,
-                        lineHeight = 40.sp,
-                        fontWeight = FontWeight.ExtraBold,
-                        color = colorResource(R.color.color_digital_blue),
-                    ),
-                )
-                Spacer(modifier = Modifier.height(46.dp))
-                val annotatedText = annotatedStringResource(
-                    id = R.string.root_description,
-                    spanStyles = { annotation ->
-                        when (annotation.key) {
-                            "is_url" -> SpanStyle(color = Color.Blue)
-                            else -> null
-                        }
-                    },
-                )
-                ClickableText(
-                    text = annotatedText,
-                    modifier = Modifier.padding(horizontal = 16.dp),
-                    style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
-                    onClick = { offset ->
-                        annotatedText.getStringAnnotations(
-                            tag = "is_url",
-                            start = offset,
-                            end = offset,
-                        ).firstOrNull()?.let {
-                            uriHandler.openUri("https://portal.smileidentity.com/sdk")
-                        }
-                    },
-                )
-            },
-            pinnedContent = {
-                Button(
-                    onClick = {
-                        scanner.startScan()
-                            .addOnSuccessListener { barcode ->
-                                barcode.rawValue?.let { config ->
-                                    viewModel.updateSmileConfig(updatedConfig = config)
-                                    isFirstTime = true
-                                }
-                            }
-                    },
-                    modifier = Modifier.padding(horizontal = 16.dp),
-                ) {
-                    Text(
-                        text = stringResource(id = R.string.root_scan_qr),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(8.dp),
-                        textAlign = TextAlign.Center,
+    SmileIDTheme {
+        Surface(modifier = modifier) {
+            if (runtimeConfig != null) {
+                if (isFirstTime) {
+                    SmileConfigConfirmationScreen(
+                        partnerId = uiState.partnerId,
+                        onConfirm = {
+                            showBottomSheet = false
+                            isFirstTime = false
+                        },
                     )
                 }
-                OutlinedButton(
-                    onClick = { showBottomSheet = true },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                ) {
-                    Text(
-                        text = stringResource(R.string.root_add_config),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(8.dp),
-                        textAlign = TextAlign.Center,
-                    )
-                }
-            },
-        )
-
-        if (runtimeConfig != null) {
-            if (isFirstTime) {
-                SmileConfigConfirmationScreen(
-                    partnerId = uiState.partnerId,
-                    onConfirm = {
-                        showBottomSheet = false
-                        isFirstTime = false
-                    },
-                )
-            } else {
                 // If a config has been set at runtime, it takes first priority
                 LaunchedEffect(runtimeConfig) {
                     initialized = false
+                    isFirstTime = false
                     SmileID.initialize(
                         context = context,
                         config = runtimeConfig!!,
@@ -200,41 +89,61 @@ private fun Content(
                     )
                     initialized = true
                 }
-            }
-        } else if (context.isConfigDefineInAssets()) {
-            // Otherwise, fallback to the config defined in assets
-            LaunchedEffect(Unit) {
-                initialized = false
-                SmileID.initialize(
-                    context = context,
-                    useSandbox = true,
-                    enableCrashReporting = !BuildConfig.DEBUG,
-                    okHttpClient = client,
-                )
-                initialized = true
-            }
-        }
-
-        if (showBottomSheet) {
-            SmileConfigModalBottomSheet(
-                onSaveSmileConfig = { config ->
-                    viewModel.updateSmileConfig(updatedConfig = config)
-                    isFirstTime = true
-                },
-                onDismiss = { showBottomSheet = false },
-                errorMessage = uiState.smileConfigError,
-                hint = uiState.smileConfigHint,
-                dismissable = true,
-            )
-        }
-
-        key(runtimeConfig) {
-            if (initialized) {
-                MainScreen()
+            } else if (context.isConfigDefineInAssets()) {
+                // Otherwise, fallback to the config defined in assets
                 LaunchedEffect(Unit) {
-                    if (!context.isInternetAvailable()) {
-                        context.toast(R.string.warning_no_internet)
+                    initialized = false
+                    isFirstTime = false
+                    SmileID.initialize(
+                        context = context,
+                        useSandbox = true,
+                        enableCrashReporting = !BuildConfig.DEBUG,
+                        okHttpClient = client,
+                    )
+                    initialized = true
+                }
+            } else {
+                isFirstTime = true
+                initialized = false
+            }
+
+            if (showBottomSheet) {
+                SmileConfigModalBottomSheet(
+                    onSaveSmileConfig = { config ->
+                        viewModel.updateSmileConfig(updatedConfig = config)
+                        isFirstTime = true
+                    },
+                    onDismiss = { showBottomSheet = false },
+                    errorMessage = uiState.smileConfigError,
+                    hint = uiState.smileConfigHint,
+                    dismissable = true,
+                )
+            }
+
+            key(runtimeConfig) {
+                if (initialized) {
+                    MainScreen()
+                    LaunchedEffect(Unit) {
+                        if (!context.isInternetAvailable()) {
+                            context.toast(R.string.warning_no_internet)
+                        }
                     }
+                } else {
+                    WelcomeScreen(
+                        showBottomSheet = {
+                            showBottomSheet = true
+                        },
+                        showQRScanner = {
+                            scanner.startScan()
+                                .addOnSuccessListener { barcode ->
+                                    barcode.rawValue?.let { config ->
+                                        viewModel.updateSmileConfig(updatedConfig = config)
+                                        isFirstTime = true
+                                    }
+                                }
+                        },
+                        modifier = Modifier.fillMaxSize(),
+                    )
                 }
             }
         }
